requirements.txt
Django==4.0.3
Pillow
numpy==1.26.3
cffi==1.16.0 
cryptography==42.0.2 
pycparser==2.21
psycopg2==2-2.9.9
jpype1==1.5.0